* develop: (22 commits)
  Update user Token API to use caches.
  Update README to clarify install instructions (closes #24, adds directory instruction referenced in #25)
  Allow a fallback token based authentication for browsers which can't read the CSRF protection cookie
  BS-136: Allow for a setting to disable account creation
  BS-878: Avoid 500 error when unauthenticated user requests /v1/issuer/all-badges. Update all issuer API calls to require authentication.
  Add system-specific requirements to README.
  #21 Alter setup instructions to use https git clone as friendly default for users who don't have SSH keys configured.
  #20, #17, #21: Fix erroneous inaccessible requirement.
  BS-837 Dropzone: image is not centered BS-836 Dropzone: background icon remains after image is dropped
  BS-850 Added borders to standard table header
  BS-846 Added min-height for badgr and collection modules
  Update README for v1.1.2 install instructions.
  BS-870 Image request for issuer image should return 404 instead of 500 when there is no (optional) image associated with the issuer.
  Update README for v1.1.2 install instructions.
  BS-839 Added modules and pattern library examples for auth views
  BS-836 BS-837 Updated pattern library to have examples of dropzone module
  only use badgebook if present
  remove badgebook
  setting next development version: 1.1.2
  Updated badge studio in pattern library to include color picker, added dialog example with nested detail and step modules
  ...

+++ apps/issuer/api.py
@@ -1,5 +1,6 @@
 from itertools import chain
 import logging
+from django.apps import apps
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -10,7 +11,6 @@ from rest_framework import status, authentication, permissions
 from rest_framework.exceptions import ValidationError, PermissionDenied
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from badgebook.models import BadgeObjectiveAward, LmsCourseInfo
 import badgrlog
 
 from .models import Issuer, IssuerStaff, BadgeClass, BadgeInstance
@@ -154,7 +154,7 @@ class IssuerDetail(AbstractIssuerAPIEndpoint):
     queryset = Issuer.objects.all()
     model = Issuer
     serializer_class = IssuerSerializer
-    permission_classes = (IsStaff,)
+    permission_classes = (permissions.IsAuthenticated, IsStaff,)
 
     def get(self, request, slug):
         """
@@ -204,10 +204,15 @@ class IssuerDetail(AbstractIssuerAPIEndpoint):
         # remove issuer staff
         IssuerStaff.objects.filter(issuer=issuer).delete()
 
-        # update LmsCourseInfo's that were using this issuer as the default_issuer
-        for course_info in LmsCourseInfo.objects.filter(default_issuer=issuer):
-            course_info.default_issuer = None
-            course_info.save()
+        if apps.is_installed('badgebook'):
+            try:
+                from badgebook.models import LmsCourseInfo
+                # update LmsCourseInfo's that were using this issuer as the default_issuer
+                for course_info in LmsCourseInfo.objects.filter(default_issuer=issuer):
+                    course_info.default_issuer = None
+                    course_info.save()
+            except ImportError:
+                pass
 
         # delete the issuer
         issuer.delete()
@@ -219,7 +224,7 @@ class IssuerStaffList(AbstractIssuerAPIEndpoint):
     role = 'staff'
     queryset = Issuer.objects.all()
     model = Issuer
-    permission_classes = (IsOwnerOrStaff,)
+    permission_classes = (permissions.IsAuthenticated, IsOwnerOrStaff,)
 
     def get(self, request, slug):
         """
@@ -339,7 +344,7 @@ class AllBadgeClassesList(AbstractIssuerAPIEndpoint):
     """
     queryset = Issuer.objects.all()
     model = Issuer
-    permission_classes = (IsEditor,)
+    permission_classes = (permissions.IsAuthenticated, IsEditor,)
 
     def get(self, request):
         """
@@ -364,7 +369,7 @@ class BadgeClassList(AbstractIssuerAPIEndpoint):
     """
     queryset = Issuer.objects.all()
     model = Issuer
-    permission_classes = (IsEditor,)
+    permission_classes = (permissions.IsAuthenticated, IsEditor,)
 
     def get(self, request, issuerSlug):
         """
@@ -455,7 +460,7 @@ class BadgeClassDetail(AbstractIssuerAPIEndpoint):
     """
     queryset = BadgeClass.objects.all()
     model = BadgeClass
-    permission_classes = (MayEditBadgeClass,)
+    permission_classes = (permissions.IsAuthenticated, MayEditBadgeClass,)
 
     def get(self, request, issuerSlug, badgeSlug):
         """
@@ -512,7 +517,7 @@ class BadgeInstanceList(AbstractIssuerAPIEndpoint):
     queryset = BadgeClass.objects.all()
     model = BadgeClass
     serializer_class = BadgeInstanceSerializer
-    permission_classes = (MayIssueBadgeClass,)
+    permission_classes = (permissions.IsAuthenticated, MayIssueBadgeClass,)
 
     def post(self, request, issuerSlug, badgeSlug):
         """
@@ -589,7 +594,7 @@ class IssuerBadgeInstanceList(AbstractIssuerAPIEndpoint):
     """
     queryset = Issuer.objects.all().select_related('badgeinstances')
     model = Issuer
-    permission_classes = (IsStaff,)
+    permission_classes = (permissions.IsAuthenticated, IsStaff,)
 
     def get(self, request, issuerSlug):
         """
@@ -633,7 +638,7 @@ class BadgeInstanceDetail(AbstractIssuerAPIEndpoint):
     """
     queryset = BadgeInstance.objects.all()
     model = BadgeInstance
-    permission_classes = (MayEditBadgeClass,)
+    permission_classes = (permissions.IsAuthenticated, MayEditBadgeClass,)
 
     def get(self, request, issuerSlug, badgeSlug, assertionSlug):
         """
@@ -685,12 +690,17 @@ class BadgeInstanceDetail(AbstractIssuerAPIEndpoint):
         current_assertion.image.delete()
         current_assertion.save()
 
-        try:
-            award = BadgeObjectiveAward.cached.get(badge_instance_id=current_assertion.id)
-        except BadgeObjectiveAward.DoesNotExist:
-            pass
-        else:
-            award.delete()
+        if apps.is_installed('badgebook'):
+            try:
+                from badgebook.models import BadgeObjectiveAward, LmsCourseInfo
+                try:
+                    award = BadgeObjectiveAward.cached.get(badge_instance_id=current_assertion.id)
+                except BadgeObjectiveAward.DoesNotExist:
+                    pass
+                else:
+                    award.delete()
+            except ImportError:
+                pass
 
         logger.event(badgrlog.BadgeAssertionRevokedEvent(current_assertion, request.user))
         return Response(
